CON-4698 added phpdoc for arg capture
